Move enter to servant package
