+++ servant-client/src/Servant/Client.hs
@@ -43,6 +43,7 @@ import           Servant.Client.Experimental.Auth
 import           Servant.Common.BaseUrl
 import           Servant.Common.BasicAuth
 import           Servant.Common.Req
+import           Servant.Utils.Enter
 
 -- * Accessing APIs as a Client
