+++ servant/src/Servant/Utils/Enter.hs
@@ -8,7 +8,7 @@
 {-# LANGUAGE TypeFamilies           #-}
 {-# LANGUAGE TypeOperators          #-}
 {-# LANGUAGE UndecidableInstances   #-}
-module Servant.Server.Internal.Enter where
+module Servant.Utils.Enter where
 
 import qualified Control.Category            as C
 #if MIN_VERSION_mtl(2,2,1)
